Enhance browser testing framework and improve type safety
- Updated `analysis_options.yaml` to include new linter rules and error handling.
- Modified `pubspec.yaml` to adjust dependencies for better compatibility.
- Introduced `status_message.dart` for centralized HTTP status message handling.
- Refactored `response.dart` to utilize the new status message utility and improve type safety.
- Enhanced various browser-related classes with better type checks and documentation.
- Improved logging functionality in `BrowserLogger` for better test traceability.
- Updated tests to reflect changes in the browser testing framework.

+++ packages/server_testing/lib/src/transport/server.dart
@@ -26,7 +26,7 @@ class ServerTransport extends TestTransport {
     final receivePort = ReceivePort();
     _isolate = await Isolate.spawn(
         _serverRunner, _ServerConfig(_handler, receivePort.sendPort));
-    _port = await receivePort.first;
+    _port = await receivePort.first as int;
   }
 
   /// Sends an HTTP request to the server and returns the response.
